test/web-platform-tests/to-upstream/css/css-cascade/layers-specificity.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>CSS Cascade: selector specificity</title>
<link rel="help" href="https://www.w3.org/TR/css-cascade-5/#cascading">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  #foo>#test1 {
    text-align: left;
  }

  #test1 {
    text-align: right;
  }

  #test2.center {
    text-align: center;
  }

  #test2 {
    text-align: left;
  }

  .test3--left.test3--right {
    text-align: right;
  }

  .test3--left {
    text-align: left;
  }

  .test4--left {
    text-align: left;
  }

  .test4--right {
    text-align: right;
  }

  #test5[data-align="right"] {
    text-align: right;
  }

  #test5 {
    text-align: left;
  }

  #test6.specific {
    text-align: left;
  }

  #test6.very.specific,
  #test6 {
    text-align: right;
  }

  /* element type vs class: class (0,1,0) > element (0,0,1) */
  span {
    text-align: left;
  }

  .test7 {
    text-align: right;
  }

  /* :not() specificity equals its argument */
  div:not(.test8-exclude) {
    text-align: right;
  }

  .test8 {
    text-align: left;
  }

  /* universal selector has zero specificity */
  * {
    text-align: left;
  }

  .test9 {
    text-align: right;
  }
</style>

<body>
  <div id="foo">
    <div id="test1"></div>
  </div>
  <div id="test2" class="center"></div>
  <div id="test3" class="test3--left test3--right"></div>
  <div id="test4" class="test4--left test4--right"></div>
  <div id="test5" data-align="right"></div>
  <div id="test6" class="very specific"></div>
  <span id="test7" class="test7"></span>
  <div id="test8" class="test8"></div>
  <div id="test9" class="test9"></div>
</body>

<script>
  "use strict";

  test(() => {
    const element = document.querySelector("#test1");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "left"
    );
  }, "2 ids vs 1 id selector");

  test(() => {
    const element = document.querySelector("#test2");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "center"
    );
  }, "1 id vs 1 id + classname");

  test(() => {
    const element = document.querySelector("#test3");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "right"
    );
  }, "1 vs 2 classnames");

  test(() => {
    const element = document.querySelector("#test4");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "right"
    );
  }, "1 vs 1 classnames");

  test(() => {
    const element = document.querySelector("#test5");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "right"
    );
  }, "id vs attrs");

  test(() => {
    const element = document.querySelector("#test6");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "right"
    );
  }, "highest specificity of matching");

  test(() => {
    const element = document.querySelector("#test7");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "right"
    );
  }, "class selector beats element type selector");

  test(() => {
    const element = document.querySelector("#test8");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "right"
    );
  }, ":not() specificity equals its argument");

  test(() => {
    const element = document.querySelector("#test9");
    assert_equals(
      getComputedStyle(element).getPropertyValue("text-align"),
      "right"
    );
  }, "universal selector has zero specificity");
</script>
